requirements.txt
streamlit>=1.35.0
ollama>=0.3.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
python-pptx>=0.6.23
watchdog>=4.0.0
duckduckgo-search>=6.0.0
pyyaml>=6.0.1
jinja2>=3.1.0
langchain>=0.2.0
langchain-community>=0.2.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0
pandas>=2.2.0
beautifulsoup4>=4.12.0
ebooklib>=0.18
